moved fixture to conftest, updated inputs to test item and user deletion from share
