+++ tests/test_add.py
@@ -3,20 +3,6 @@ import pytest
 import os
 import subprocess
 
-
-@pytest.fixture(scope='function')
-def single_file_share(source_dir, shares_dir, calling_user, calling_prim_group, variables):
-    from nfs4_share.manage import create
-    items = fabricate_a_source(source_dir, [
-        "file"
-    ])
-    share = create(shares_dir.join('share'), items=items, users=[calling_user], managing_groups=[calling_prim_group],
-                   user_apache_directive=variables["user_directive"], group_apache_directive=variables["group_directive"],
-                   domain=variables["domain_name"],
-                   service_application_accounts=variables['service_application_accounts'])
-    return share
-
-
 def test_null_add_share(single_file_share):
     from nfs4_share.manage import add
     add(single_file_share.directory)
